Set authz checked when rate limiting device endpoints
